- dashboard and single job blocks new structure
