--- includes/blocks/jb-job/block.asset.php
@@ -1,10 +1,0 @@
-<?php return array(
-	'dependencies' => array(
-		'wp-blocks',
-		'wp-element',
-		'wp-components',
-		'wp-server-side-render',
-		'wp-polyfill',
-	),
-	'version'      => JB_VERSION,
-);
