+++ package.json
@@ -9,15 +9,18 @@
     "email": "team@ultimatemember.com",
     "url": "https://ultimatemember.com"
   },
-  "dependencies": {},
   "devDependencies": {
+    "@wordpress/block-editor": "^11.5.0",
+    "@wordpress/blocks": "^12.5.0",
+    "@wordpress/components": "^23.5.0",
+    "@wordpress/scripts": "^25.5.0",
     "gulp": "^4.0.2",
     "gulp-clean-css": "^4.3.0",
     "gulp-cli": "^2.3.0",
     "gulp-rename": "^2.0.0",
-    "gulp-sass": "^4.1.0",
+    "gulp-sass": "^5.1.0",
     "gulp-uglify": "^3.0.2",
-    "jsdoc": "~3.6.3",
+    "jsdoc": "~4.0.2",
     "wp-hookdoc": "^0.2.0"
   },
   "scripts": {
